% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSquared.R
\name{ChiSquared}
\alias{ChiSquared}
\title{Compare 2 Categorical Variables with Chi-Squared}
\usage{
ChiSquared(
  x = ProjectData$gender,
  y = ProjectData$phys,
  xname = "Gender",
  yname = "Physical Activity",
  sig.fig = 0.05
)
}
\arguments{
\item{x}{First variable data, default is ProjectData$gender}

\item{y}{Second variable data, default is ProjectData$phys}

\item{xname}{The name of the first variable, default Gender}

\item{yname}{The name of the second variable, default Physical Activity}

\item{sig.fig}{Significance level to be tested at, default 0.05 (5\%)}
}
\value{
Output list containing the hypothesis, assumptions, test statistics, decision and a conclusion of the test.
}
\description{
When passed 2 vectors for different categories the function will test if there is a difference in distribution because of the different categories.
}
\examples{
ChiSquared(x = sample(c('male', 'female'), 100, replace=TRUE),
y = sample(c('bald', 'short', 'long'), 100, replace=TRUE))
}
