% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearModel.R
\name{LinearModel}
\alias{LinearModel}
\title{Test Linear Relationship}
\usage{
LinearModel(
  X = ProjectData$height,
  Y = ProjectData$weight,
  xName = "height",
  yName = "weight",
  sig.fig = 0.05
)
}
\arguments{
\item{X}{First variable data}

\item{Y}{Second variable data}

\item{xName}{The name of the first variable}

\item{yName}{The name of the second variable}

\item{sig.fig}{Significance level to be tested at}
}
\value{
output list containing the hypothesis, assumptions, test statistics, decision and a conclusion.
}
\description{
When given 2 numerical vectors the function will test if there is a linear relationship between them. The resulting output will include a hypothesis, assumptions, test statistics, decision and a conclusion.
}
\examples{
LinearModel(seq(100, 200, 2.5), runif(41, 100, 200))
}
