% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectData.R
\docType{data}
\name{ProjectData}
\alias{ProjectData}
\title{Physical results for 1000 people}
\format{
A data frame with 1000 observations of 4 variables and each persons ID:
\describe{
\item{ID}{The identification reference for each entry.}
\item{gender}{The gender of each person, simply male of female.}
\item{height}{The height of each person.}
\item{weight}{The height of each person.}
\item{phys}{The level of physical activity of each person, levels are none, moderate, and intense.}

}
}
\usage{
ProjectData
}
\description{
A dataset containing the gender, height, weight, and physical activity for 1000 different people
}
\keyword{datasets}
