% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectFemale.R
\docType{data}
\name{ProjectFemale}
\alias{ProjectFemale}
\title{Physical results of FeFemales}
\format{
A data frame with 507 observations of 4 variables and each persons ID:
\describe{
\item{ID}{The identification reference for each entry.}
\item{gender}{The gender of each person, Female.}
\item{height}{The height of each person.}
\item{weight}{The height of each person.}
\item{phys}{The level of physical activity of each person. Levels are none, moderate, and intense.}

}
}
\usage{
ProjectFemale
}
\description{
A dataset, based on ProjectData, with only Females selected. Dataset contains each womans height, weight, and physical activity for 1000 different people
}
\keyword{datasets}
