% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectMale.R
\docType{data}
\name{ProjectMale}
\alias{ProjectMale}
\title{Physical results of Males}
\format{
A data frame with 507 observations of 4 variables and each persons ID:
\describe{
\item{ID}{The identification reference for each entry.}
\item{gender}{The gender of each person, male.}
\item{height}{The height of each person.}
\item{weight}{The height of each person.}
\item{phys}{The level of physical activity of each person. Levels are none, moderate, and intense.}

}
}
\usage{
ProjectMale
}
\description{
A dataset, based on ProjectData, with only Males selected. Dataset contains each mans height, weight, and physical activity for 1000 different people
}
\keyword{datasets}
