% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_report.R
\name{ttest_report}
\alias{ttest_report}
\title{Test 2 sample means}
\usage{
ttest_report(
  x = ProjectMale$height,
  y = ProjectFemale$height,
  cata = "height",
  xname = "Male",
  yname = "Female",
  sig.fig = 0.05
)
}
\arguments{
\item{x}{First variable data, default is ProjectMale$height}

\item{y}{Second variable data, default is ProjectFemale$height}

\item{cata}{The category for each variable to be tested, default is height}

\item{xname}{The name of the first variable, default Male}

\item{yname}{The name of the second variable, default Female}

\item{sig.fig}{Significance level to be tested at, default 0.05 (5\%)}
}
\value{
Output list containing the hypothesis, assumptions, test statistics, decision and a conclusion of the test.
}
\description{
When given 2 numerical vectors the function will test if there is a statistically significant difference between each of their means.
}
\examples{
ttest_report(rpois(20, 10), rpois(20, 11), cata = "teeth")
}
